bring cli tests together (api online and offline)
